+++ components/common/Button.tsx
@@ -1,22 +1,33 @@
+import { ReactNode } from "react";
 import cn from "classnames";
 
 type ButtonProps = {
     onClick: React.MouseEventHandler<HTMLButtonElement>;
-    label: string;
+    label: string | ReactNode;
+    name?: string;
     additionalClasses?: string;
-    variant?: "default" | "light" | "dark" | "green" | "red";
+    variant?: "base" | "round" | "quantity";
+    theme?: "default" | "light" | "dark" | "green" | "red";
 } & Omit<React.ButtonHTMLAttributes<HTMLButtonElement>, "onClick">;
 
-const base = `
-    font-semibold py-2 px-4 rounded shadow transition-colors border cursor-pointer
-`;
+const defaultClasses = "font-semibold transition-colors border cursor-pointer"
+
+const variants = {
+    base: `
+        py-2 px-4 rounded shadow"
+    `,
+    round: `
+        p-2 bg-white rounded-full shadow-lg hover:bg-gray-100
+    `,
+    quantity: `
+        w-7 h-7 flex justify-center items-center hover:bg-neutral-600 hover:text-white
+    `
+};
 
 const themes = {
     default: `
         text-gray-800 hover:text-gray-900 
-        dark:text-white dark:hover:text-gray-100
-        border-gray-400 dark:border-gray-600
-        hover:border-gray-500 dark:hover:border-gray-300
+        border-gray-400 hover:border-gray-500 
     `,
     light: `
         text-gray-800 hover:text-gray-900
@@ -32,30 +43,30 @@ const themes = {
         text-white
         bg-green-600 hover:bg-green-700
         border-green-700 hover:border-green-800
-        dark:bg-green-500 dark:hover:bg-green-600
-        dark:border-green-600 dark:hover:border-green-700
     `,
     red: `
         text-white
         bg-red-600 hover:bg-red-700
         border-red-700 hover:border-red-800
-        dark:bg-red-500 dark:hover:bg-red-600
-        dark:border-red-600 dark:hover:border-red-700
     `,
 };
 
 export default function Button({
     onClick,
     label,
+    name,
     additionalClasses,
-    variant = "default"
+    variant = "base",
+    theme = "default",
 }: ButtonProps) {
     return (
         <button
             onClick={onClick}
+            name={name}
             className={cn(
-                base,
-                themes[variant],
+                defaultClasses,
+                variants[variant],
+                themes[theme],
                 additionalClasses
             )}
         >