+++ components/common/Container.tsx
@@ -3,15 +3,19 @@ import cn from "classnames";
 
 type ContainerProps = {
     children: ReactNode;
+    containerSize?: "lg" | "xl";
     additionalClasses?: string;
 };
 
-export default function Container({ children, additionalClasses }: ContainerProps) {
+const containerWidths = {
+    "lg": "max-w-lg",
+    "xl": "max-w-[1500px]"
+}
+
+export default function Container({ children, containerSize = "xl", additionalClasses }: ContainerProps) {
     return (
         <div
-            className={cn("max-w-[1500px] mx-auto px-8", {
-                additionalClasses
-            })}
+            className={cn("mx-auto px-8", containerWidths[containerSize], additionalClasses)}
         >
             {children}
         </div>